requirements.txt
Flask
Pillow
Werkzeug
tensorflow
opencv-python
numpy
